+++ Sources/SuperwallKit/Dependencies/DependencyContainer.swift
@@ -40,6 +40,7 @@ final class DependencyContainer {
   var attributionPoster: AttributionPoster!
   var webEntitlementRedeemer: WebEntitlementRedeemer!
   var deepLinkRouter: DeepLinkRouter!
+  var attributionFetcher: AttributionFetcher!
   // swiftlint:enable implicitly_unwrapped_optional
   let paywallArchiveManager = PaywallArchiveManager()
 
@@ -115,10 +116,17 @@ final class DependencyContainer {
       factory: self
     )
 
+    attributionFetcher = AttributionFetcher(
+      storage: storage,
+      deviceHelper: deviceHelper,
+      webEntitlementRedeemer: webEntitlementRedeemer
+    )
+
     attributionPoster = AttributionPoster(
       storage: storage,
       network: network,
-      configManager: configManager
+      configManager: configManager,
+      attributionFetcher: attributionFetcher
     )
 
     placementsQueue = PlacementsQueue(
@@ -520,8 +528,8 @@ extension DependencyContainer: PurchasedTransactionsFactory {
 
 // MARK: - User Attributes Placement Factory
 extension DependencyContainer: UserAttributesPlacementFactory {
-  func makeUserAttributesPlacement() -> InternalSuperwallEvent.Attributes {
-    return InternalSuperwallEvent.Attributes(
+  func makeUserAttributesPlacement() -> InternalSuperwallEvent.UserAttributes {
+    return InternalSuperwallEvent.UserAttributes(
       appInstalledAtString: deviceHelper.appInstalledAtString,
       audienceFilterParams: identityManager.userAttributes
     )